refactor: rename root component files
