+++ packages/spectacular/src/index.ts
@@ -20,6 +20,4 @@ export * from './lib/pipe-test-harness/create-pipe-test-harness-options';
 export * from './lib/pipe-test-harness/pipe-test-harness';
 
 // Root component
-export * from './lib/root-component/spectacular-app.module';
-export * from './lib/root-component/spectacular-root.component';
-export * from './lib/root-component/spectacular-root.scam';
+export * from './lib/root-component/spectacular-app.component';
