+++ packages/spectacular/src/lib/feature-testing/testbed/spectacular-feature-testbed.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { Router } from '@angular/router';
 
-import { SpectacularAppComponent } from '../../root-component/spectacular-root.component';
+import { SpectacularAppComponent } from '../../root-component/spectacular-app.component';
 import { SpectacularFeatureTestingModule } from '../feature-testing-module/spectacular-feature-testing.module';
 import { CreateFeatureOptions } from './create-feature-options';
