+++ packages/spectacular/src/lib/root-component/spectacular-app.component.ts
